Style components of the main game screen and add TODO/FIXME comments
This commit styles the majority of the components used to display the main game screen with a filler image for the villager cards and filler text for the villager names. These fillers will be replaced once a method for fetching and using data from an API call is implemented in the future.

This commit also adds TODO/FIXME comments to keep track of future project goals.