Fix background image and responsiveness of card layouts on different sized screens
This commit fixes the background image by setting its size to 'contain' the image and prevent unwanted repeating on different sized screens.

This commit also changes the layout of the cards based on different sized screens using media queries. This way, there is an even number of cards per column/row, which the previous code did not do.