Update the round number and scoreboard after each round and complete UI
This commit updates the scoreboard and round number by checking whether or not the selected card has already been clicked or not. If the card has been clicked, the game ends and a loss message appears, prompting the user to continue playing with the same cards but are reset. If the card has not been clicked, however, the round number and scoreboard updates and a max score of 8 displays a win message that gives the user the option to either refresh the page or click on any card to reset the game with new cards.